Thiago

+++ app/Http/Controllers/ProdutoController.php
@@ -5,15 +5,18 @@ namespace App\Http\Controllers;
 use App\Models\Categoria;
 use App\Models\Produto;
 use App\Models\Estoque;
+use App\Models\Compra;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Auth;
 
-class ProdutoController extends Controller {
+class ProdutoController extends Controller
+{
     /**
      * Display a listing of the resource.
      */
-    public function index() {
+    public function index()
+    {
         $produtos = Produto::all();
         return view('adm.produto.index', compact('produtos'));
     }
@@ -29,7 +32,8 @@ class ProdutoController extends Controller {
     /**
      * Show the form for creating a new resource.
      */
-    public function create() {
+    public function create()
+    {
         $categorias = Categoria::all();
         return view('adm.produto.create', compact('categorias'));
     }
@@ -58,32 +62,48 @@ class ProdutoController extends Controller {
     /**
      * Display the specified resource.
      */
-    public function show(Produto $produto) {
+    public function show(Produto $produto)
+    {
         return view('adm.produto.show', compact('produto'));
     }
 
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(Produto $produto) {
-        return view('adm.produto.edit');
+    public function edit($id_produto)
+    {
+        $produto = Produto::findOrFail($id_produto);
+        $categorias = Categoria::all();
+        return view('adm.produto.edit', compact('produto', 'categorias'));
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, Produto $produto)
+    public function update(Request $request, $id_produto)
     {
-        //
+        $produto = Produto::findOrFail($id_produto);
+        if($request->method() == 'GET'){
+            return view('adm.produto.edit', compact('produto'));
+        } else {
+            $produto->update($request->all());
+        }    
+        return redirect()->route('adm.produto.index')->with('message', 'Produto atualizado com sucesso!');
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Produto $produto)
+    public function destroy($id_produto)
     {
-        $produto = Produto::findOrFail($produto->id_produto);
-        $produto->delete();
-        return redirect()->route('adm.produto.index')->with('message','Deletado com sucesso!');
+        Compra::where('id_produto', $id_produto)->delete();
+
+        $produto = Produto::find($id_produto);
+        if ($produto) {
+            $produto->delete();
+            return redirect()->route('adm.produto.index')->with('message', 'Deletado com sucesso!');
+        } else {
+            return redirect()->route('adm.produto.index')->with('error', 'Produto não encontrado!');
+        }
     }
 }
